Move all tests to tests directory.
Add separate settings file (tests/settings.py).
Refactor import according to the new structure.
