+++ foreman/tests/foreman/management/commands/test_correct_affy_cdfs.py
