+++ foreman/tests/foreman/management/commands/test_create_compendia.py
