+++ foreman/tests/foreman/management/commands/test_update_experiment_metadata.py
@@ -1,5 +1,3 @@
-from unittest.mock import patch
-
 from django.test import TransactionTestCase
 
 from data_refinery_common.models import Experiment, ExperimentSampleAssociation, Sample
