+++ foreman/tests/foreman/test_job_control.py
