+++ foreman/tests/settings.py
@@ -1,0 +1,5 @@
+"""
+Django settings for data_refinery_foreman project tests.
+"""
+
+from data_refinery_foreman.settings import *
