+++ foreman/tests/test_runner.py
@@ -2,12 +2,12 @@ from django.test.runner import DiscoverRunner
 
 
 class NoDbTestRunner(DiscoverRunner):
-    """ A test runner to test without database creation """
+    """A test runner to test without database creation"""
 
     def setup_databases(self, **kwargs):
-        """ Override the database creation defined in parent class """
+        """Override the database creation defined in parent class"""
         pass
 
     def teardown_databases(self, old_config, **kwargs):
-        """ Override the database teardown defined in parent class """
+        """Override the database teardown defined in parent class"""
         pass
